+++ src/features/calendar/widgetCalendar/components/ScheduleListItem.tsx
@@ -3,12 +3,12 @@ import React from 'react';
 import { View, Text } from 'react-native';
 import Animated, { FadeIn } from 'react-native-reanimated';
 
+import { Category } from '@/shared/types/categoryType';
+
 interface ScheduleListItemProps {
-  category: string;
+  category: Category;
   detail: string;
   time: string;
-  categoryBackgroundColor: string;
-  categoryTextColor: string;
   memo: string;
 }
 
@@ -16,8 +16,6 @@ const ScheduleListItem = ({
   category,
   detail,
   time,
-  categoryBackgroundColor,
-  categoryTextColor,
   memo,
 }: ScheduleListItemProps) => {
   return (
@@ -26,11 +24,11 @@ const ScheduleListItem = ({
         <View className="flex-1 flex-row items-center space-x-2">
           <View
             className="rounded-full px-[11px] py-1"
-            style={{ backgroundColor: categoryBackgroundColor }}>
+            style={{ backgroundColor: category.backgroundColor }}>
             <Text
               className="text-white typo-caption-14-light"
-              style={{ color: categoryTextColor }}>
-              {category}
+              style={{ color: category.textColor }}>
+              {category.categoryName}
             </Text>
           </View>
           <Text
@@ -45,7 +43,12 @@ const ScheduleListItem = ({
         </Text>
       </View>
       <View className="mt-[9px] rounded-[15px] bg-surface-100 px-[16px] py-[14px]">
-        <Text className="text-main-text typo-caption-13-medium">{memo}</Text>
+        <Text
+          className="text-main-text typo-caption-13-medium"
+          numberOfLines={1}
+          ellipsizeMode="tail">
+          {memo}
+        </Text>
       </View>
     </Animated.View>
   );
