+++ src/features/calendar/widgetCalendar/hooks/useMiniCalendarAnimation.ts
